requirements.txt
Flask==2.3.2
Flask-SQLAlchemy==2.5.1
Flask-Login==0.5.0
googlemaps~=4.10.0
requests~=2.32.3
python-dotenv~=1.0.1